move extensions specs to extensions spec file
